+++ revdep/library.noindex/greta/new/Rcpp/include/Rcpp/Function.h
@@ -82,14 +82,10 @@ namespace Rcpp{
             return Rcpp_fast_eval(call, R_GlobalEnv);
         }
 
-        #if defined(HAS_VARIADIC_TEMPLATES) || defined(RCPP_USING_CXX11)
-            template <typename... T>
-            SEXP operator()(const T&... args) const {
+        template <typename... T>
+        SEXP operator()(const T&... args) const {
             return invoke(pairlist(args...), R_GlobalEnv);
-            }
-        #else
-            #include <Rcpp/generated/Function__operator.h>
-        #endif
+        }
 
         /**
          * Returns the environment of this function
@@ -99,7 +95,11 @@ namespace Rcpp{
             if( TYPEOF(fun) != CLOSXP ) {
                 throw not_a_closure(Rf_type2char(TYPEOF(fun)));
             }
-            return CLOENV(fun) ;
+            #if (defined(R_VERSION) && R_VERSION >= R_Version(4,5,0))
+            return R_ClosureEnv(fun);
+            #else
+            return CLOENV(fun);
+            #endif
         }
 
         /**